+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features
 * feat: add offline mode toast
 * feat: add offline mode page
+* add offline mode bundler
 
 ## Version 1.1.0 (2022-06-14)
 ### Features